requirements.txt
asgiref==3.7.2
Django==4.2.4
django-appconf==1.0.5
django-avatar==7.1.1
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
